Corrected a bug in the process_data.py and visualize_data.py modules, where the normalization of the images was done wrong. The initial assumption was that all the images have the same size, but that was modified. Also converted the labels from the coco_paper format into the coco2017 format used by Yolo.
Created the pre_trained.py module, which runs the inference of the data on an existing pretrained model, in this case it uses the YoloV8m model. The module runs the inference on the first 100 test images, and it can show the predicted bbox, that is in green, the annotations that are in blue, and the intersection between those, those are in red. The module also computes some basic statistics of the model like class accuracy, global accuracy, and the average iou value per class.